+++ Samples~/Quickstart/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("EmbodiedLab.Unity.Samples.Quickstart.Imported.Tests")]